requirements.txt
cassandra-driver
colbert-ai
torch
